Play 2 tortoises on Playground Book
