+++ PlaygroundBook/Tortoise.playground/contents.xcplayground
@@ -1,2 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='6.0' target-platform='ios' display-mode='raw'/>
+<playground version='6.0' target-platform='ios' display-mode='raw'>
+    <pages>
+        <page name='Page1'/>
+        <page name='Page2'/>
+    </pages>
+</playground>
