+++ PlaygroundBook/template/Contents/Chapters/Main.playgroundchapter/Pages/Page2.playgroundpage/Manifest.plist
@@ -5,7 +5,7 @@
 	<key>Version</key>
 	<string>1.0</string>
 	<key>Name</key>
-	<string>🐢 Playground</string>
+	<string>Play with 🐢🐇</string>
 	<key>LiveViewMode</key>
 	<string>VisibleByDefault</string>
 	<key>LiveViewEdgeToEdge</key>
